pyproject.toml
[project]
name = "assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.9.3",
    "numpy>=2.1.3",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "rich>=13.9.4",
    "scikit-image>=0.24.0",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "torch-rim>=0.2.9",
    "torchvision>=0.20.1",
    "treescope>=0.1.6",
    "utils",
    "wandb>=0.19.1",
]

[tool.uv.workspace]
members = ["utils"]

[tool.uv.sources]
utils = { workspace = true }
